refactor: place classes related to exceptions in a folder
