+++ app/src/main/java/ru/itis/androidhomework/presentation/screens/MainList/MainListFragment.kt
@@ -1,6 +1,6 @@
 package ru.itis.androidhomework.presentation.screens.MainList
 
-import android.os.Bundle
+import  android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
